pyproject.toml
[project]
name = "migrate-tel-to-spoti"
version = "0.1.0"
description = "Migrate your music playlist from a telegram channel to a spotify playlist"
authors = [
    {name = "Amin Gharibi",email = "Gharibi8364@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "telethon (>=1.39.0,<2.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "spotipy (>=2.25.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
